I can now ctrl-c the solver within python
Python is silly. There's some nuance about signal handling where it sets a
SIGINT (ctrl-c) handler to just set a flag, and the python layer then reads this
flag and does the thing. Here I'm running C code, so SIGINT would set a flag,
but not quit, so I can't interrupt the solver. Thus I reset the SIGINT handler
to the default, and put it back to the python-specific version when I'm done