OpenCV distortion models supported
This is a large change because intrinsics can now have variable sizes. The
foundation for supporting CAHVOR is laid.

There was what looks to be a bug in OpenCV where the cvProjectPoints2()
gradients of the intrinsics couldn't be non-contiguous slices of a larger array
or something. Whatever. It works now.