confidence computations use cholmod_solve() instead of cholmod_spsolve()
I'm solving JtJ x = b where J is sparse, b is sparse, but x ends up dense.
cholmod doesn't have functions for this exact case. so I use the
dense-sparse-dense function, and densify the input. Instead of
sparse-sparse-sparse and the densifying the output. This feels like it'd be more
efficient