DISTORTION_OPENCV >= 8: favor a LIGHT rational model
The radial distortion in opencv is x_distorted = x*scale where r2 = norm2(xy -
xyc) and

scale = (1 + k0 r2 + k1 r4 + k4 r6)/(1 + k5 r2 + k6 r4 + k7 r6)

Note that k2,k3 are tangential (NOT radial) distortion components. Note that the
r6 factor in the numerator is only present for >=DISTORTION_OPENCV5. Note that
the denominator is only present for >= DISTORTION_OPENCV8. The danger with a
rational model is that it's possible to get into a situation where scale ~ 0/0 ~
1. This would have very poorly behaved derivatives. If all the rational
coefficients are ~0, then the denominator is always ~1, and this problematic
case can't happen. I favor that in two ways:

- I attenuate the random seed to start closer to 0
- I regularize the rational coefficients stronger