Added C, python API to get the 'next' distortion model in a family
In a family of distortion models we have a sequence of models with increasing
complexity. Subsequent models add distortion parameters to the end of the
vector. Ealier models are identical, but with the extra paramaters set to 0.
This function returns the next model in a sequence.

If this is the last model in a sequence, returns the same model as the input
If the sequence diverges (DISTORTION_NONE is the start for all families),
returns DISTORTION_INVALID