incremental solve: I look at the distribution of NEW errors
It's possible that increasing the roi adds a SMALL NUMBER of points total. If
the model doesn't fit the new ROI well, these points would have high errors, but
since there aren't a lot of them, this wouldn't reflect at all in the new
distribution. I re-weight the distribution of the NEW points to be equal to the
distribution of all the previous points in the set