Simpler, adaptive outlier rejection
- New algorithm lives HERE, not in libdogleg

- This is both more simple and less simple. I look ONLY at the residual, without
  thinking of what happens to the other reoptimized residuals. And I don't look
  at the effect on confidences

- mrcal C API is now given the observed-pixel-uncertainty

Logic:

I define an outlier as a feature that's > k stdevs past the mean. The
threshold stdev is the worse of
- The stdev of my data set
- The expected stdev of my noise (calibrate-cameras
  --observed-pixel-uncertainty)

The rationale:

- If I have a really good solve, the stdev of my data set will be very
low, so deviations off that already-very-good solve aren't important. I
use the expected-noise stdev in this case

- If the solve isn't great, my errors may exceed the expected-noise stdev
(if my model doesn't fit very well, say). In that case I want to use
the stdev from the data