mrcal-calibrate-cameras: diagnostics are more careful about outliers, scaling
Originally, before outlier-rejection and before mrgingham reported point
confidences, and I used those to scale the measurements, I could re-compute the
reprojection errors off the state, and this would recreate the measurements
vector. So the diagnostics were cavalier about using "reprojection errors" and
"measurements" interchangeably. This patch cleans that up. The "worst-case
observations" are now sorted by the mean measurements: these are the worst
according to the optimizer. And the visualizations show the measurements, where
appropriate