get_mapping_file_framecamera(): I still parse frame numbers given one camera
So now images frame10,frame20,frame30 will give me frame numbers (10,20,30)
instead of (0,1,2). This logic was making something easier somewher, but there's
no tests or a note, so I'm removing this. If something IS broken by this, the
"fix" I removed is the wrong one: preserving frame numbers is important