show_distortion('radial') uses the correct domain
I was making a pinhole approximation in the domain of the plot produced here.
This was incorrect, and I now produce a plot based on the view angle-off-center.
It turns out many of my models had a singularity. And the distortion was far
more significant than the previous version of the tool showed