Big rework of the internal project() function
The internal function returns its intrinsics gradients in some
(lensmodel-defined) set of

- core focal-length gradients
- dense parameter gradients
- sparse parameter gradients

The Python mrcal_project is STILL returning the same dense jacobian it was
before.

It is now assumed that dq/dcxy = I and that dq/dfxy is diagonal. This saves
storing some data, and makes the big optimization jacobian sparser than it was
before.

This is a no-op patch, but it gives me flexibility to handle splined models'
gradients efficiently