major rework of the splined lens model
The previous projection function was

  u = stereographic(p)
  q = u * fxy(u, intrinsics) + cxy

This turned out to have a critical problem: at either center axis of the imager
(px=0 or py=0): the splined surface had no freedom to represent motion across
that axis. For instance, let's say I'm projecting [0,y,z]. I'd have u = [0,uy]
and q = [cx, uy*fy+cy]. Note that q=cx regardless of what the spline surface fxy
does.

This patch changes the projection function to

  u = stereographic(p)
  q = (u + deltau(u)) * fxy + cxy

So the stereographic projection is still the basis, but the value I get from the
surface is an adjustment to this stereographic projection. The fxy scaling and
cxy come from the optimization variables again. This is actually very nice: the
intrinsics core is back EXACTLY how it appears in the other models I support. I
think this requires regularization now, but that's ok: the existing code does
the right thing already.

This patch includes updates to the tests: they pass. Very promising