API for SFM fixed points changed
Instead of specifying this as a flag in
indices_point_camintrinsics_camextrinsics_flags, I now give a single
Npoints_fixed, which locks the last this many points. This is much more clear,
and works correctly: I have a flags for each observation, NOT for each point, so
was possible for a point to both be fixed and not fixed. And previously I was
including all the points in the optimization array, even the fixed ones. This
addresses all those issues