show_calibation_geometry() improvements
I can pass in a dict of point_labels, to label specific points. The rest are
plotted together as one big dataset.

The individual camera axis labels are plotted together in one big dataset, so I
can turn them ALL of together.

The camera axis labels are plotted in the correct location. This was broken for
large (SFM, mostly) problems.