suspicious points are identified correctly
I have hoaky logic used in SFM problems to identify and mitigate points that are
way too far or behind the cameras. The previous implementation wasn't looking at
the camera's geometry, which is incorrect in the general case. This patch
resolves that.

The logic is still hoaky, and should be fixed and/or removed, however