I no longer depend on the opencv C API!
Or the C++ API for that matter. My usage of opencv is now python-only.

A consequence of this patch is that LENSMODEL_OPENCV14 is no longer supported.
By I never once used this, and nobody is likely to want this ever. It can be put
back if somebody bugs me