API rework for compute_Rcorrected_dq_dintrinsics(), compute_projection_stdev()
Previously I'd pass in a bunch of observation vectors. These would be used to
compute the compensating rotation AND the same vectors would be used to evaluate
the uncertainty. Now these are two separate things, and the grid doesn't need to
be passed-in. So you can ask a question such as "what is the projection
uncertainty at this one vector v"?