Removed measurement-skipping options
These are the Python skipped_observations_board and skipped_observations_point.
And the C skip_frame, skip_observation, skip_point.

These are useful in theory, but I have NEVER used them. And there are no tests.
And the logic is really complicated. AND I'm pretty sure some of the new
uncertainty stuff won't work right if I actualy begin to use these. Tests are a
requirement for these, so I can bring these back WITH TESTS if I decide to. In
the meantime I should be able to use the outlier-rejection mechanism for these.