Reworked the outlier-rejection interface
This function updates the APIs (C and python) in incompatible ways, but makes no
functional changes.

I now use the observation weight to denote outliers: weight<0 == outlier. This
is used for both input and output. So no more "outlier_indices" arrays are
passed back/forth in C or in Python. And the observations_board_pool are no
longer const