icam_extrinsics_covariances_ief is now reported by a separate function
NOT by the optimizerCallback. C and Python APIs affected. This patch also updates
the optimization_inputs dict and the .cameramodel format. icam is no longer a
part of the optimization_inputs, but rather a separate integer.

AND this patch adds C and Python APIs to get the corresponding icam_extrinsics
to a given icam_intrinsics