Initial, VERY EARLY implementation of relative regularization for splined models
Previously I had absolute L2 regularization (pulling all the spline control
points towards 0). This maybe was causing a bias, but it also made the no-data
edges appear very confident: input data noise didn't change the projection
there, since it was 100% driven by the regularization terms.

The new approach pulls the difference between each control point and its
neighbors to 0, which hopefully would address these issues. In this
representation any planar surface has a regularization cost of 0; this maybe is
fine.

Early indications are that the new approach works, but needs more testing,
documentation and so on