splined-model regularization rework
Prior to this patch I was using an X-pattern regularization, only at internal
grid points, where all corners of the X were in-bounds. This didn't generate
enough constraints, and I was getting singular JtJ. I'm now defining the
regularization cost at ALL the grid points, even those at the edges. I simply
compute the mean among the parts of the pattern that are in-bounds. And I thus
moved back to the + pattern. This is simpler, and there's no reason to avoid it
anymore