make_synthetic_board_observations() returns more consistent random poses
This is mostly a no-op. The RNG is now sampled in one big block instead of in
two separate blocks. This makes the dance-study more consistent: if I ask for
N-1 random poses I get the same exact result as what I'd get if I ask for N
random poses and throw away the last one