implied_Rt10__from_unprojections() uses a huber loss function
Previously I was using loss='soft_l1', but it behaved strangely. For large
f_scale_deg it should be equivalent to loss='linear', but I was seeing large
diffs when comparing a model to itself:

  ./mrcal-show-projection-diff --gridn 50 28 test/data/cam0.splined.cameramodel{,} --distance 3

f_scale_deg needs to be > 0.1 to make test-projection-diff.py pass, so there was
an uncomfortably-small usable gap for f_scale_deg. loss='huber' should work
similar-ish to 'soft_l1', and it works even for high f_scale_deg