reworked the calibration-time residual reporting
Much simpler code. Fewer similar-yet-subtly-different functions to compute
stuff, and global variables to store it. This patch is at a functional place.
Subsequent patches will fix more stuff in this area.

And I removed the ability to look at errors from stuff we threw out as outliers.
It doesn't really matter, and it added completely unjustified complexity to the
code