Improved splined model regularization
I do directional L2 regularization. At each knot I penalize contributions in the
tangential direction much more than in the radial direction. Otherwise noise in
the data produces lots of curl in the vector field. This isn't wrong, but it's
much nicer if "right" in the camera coordinate system corresponds to "right" in
pixel space.

Here I should precompute the radial directions [ux,uy], but this patch is lazy,
and just computes those in the inner loop. This is much slower than it needs to
be, but at calibration time the hot path isn't here, it's in
cholmod_factorize(). So it doesn't matter when doing the slow thing. This patch
does add a #warning so that this can be fixed eventually.

The scale factors for the regularization may need to be adjusted still