More regularization when using a splined model
I need to control this better, but this is sufficient for now. I need stronger
regularization for splined models to effectively eliminate the curl in the
splined model vector field. For other models I use 2.0 because that's what I had
for a long time, and I don't want to change it to not break anything