observed_pixel_uncertainty read from residuals instead of being given
This is an API change and a conceptual change. As we get more input data, the
distribution of the residuals approaches the distribution of the input noise. I
assume that I'm not overfitting (i.e. I don't have too much data), so I use the
residual distribution as a proxy for the distribution of the input noise. THEN I
propagate that input noise for my uncertainty computations.

If I have too little data, this estimate of the input noise will be too tight,
so the uncertainty estimate will be too optimistic. So I should probably have
some sort of lower bound on this estimate of the input uncertainty. I'm not 100%
sure, but I suspect that practically speaking I'll always have enough data, and
this doesn't matter.

API changes: C code doesn't take observed_pixel_uncertainty anymore. This was a
part of the outlier rejection. It was lightly used, and was not worth the
trouble.

mrcal.optimize() and mrcal.optimizer_callback() can still take an
'observed_pixel_uncertainty' argument, but this is done purely for backwards
compatibility. This argument does nothing

m-c-c doesn't have an --observed-pixel-uncertainty argument anymore