uncertainty, diff, splined-model-corrections viz shows a zigzag
Actually, it doesn't. Disabled by default. I don't want a separate legend line
for each zigzag, so I turn this off for now.

Enabled with observations=True. This is a curve connecting the points in a board
observation. Makes it clear which points belong together in a single board
observation