mrcal-convert-lensmodel --viz: non-sampled solves recompute implied_Rt10
For a non-sampled solve, the implied_Rt10 comes from a Procrustes fit
(minimizing error in 3D space), which is usually not close-enough to get a diff.
We let the diff method figure out the implied_Rt10, and we use the uncertainties
that we have to do it.

This is potentially slow since the uncertainties need to be computed. I might
fix that later.