mrcal-image initialization reorder
This is a no-op in C. But in C++ it makes it work. Because C++ is really stupid.
They don't support C99 designated initializers (but I guess they will in
C++-20). It can sorta build, but it really cares about the initialization order.
This patch reorders things in a way it likes, and g++-12 can now #include this
stuff.