Fixed intermittent failure in test/test-projection-diff.py
I was seeing the "Low-enough diff with low focus_radius" test fail maybe 5% of
the time. Debugging showed that this is tied to the rng seed. On my machine
(vanilla amd64 with a recent Debian ) I could make it fail every time with this patch:

  diff --git a/test/test-projection-diff.py b/test/test-projection-diff.py
  index 2e58ca4..2de53a9 100755
  --- a/test/test-projection-diff.py
  +++ b/test/test-projection-diff.py
  @@ -17,6 +17,10 @@ sys.path[:0] = f"{testdir}/..",
   import mrcal
   import testutils

  +import time
  +s = int(time.time()) & 0xffff
  +s = 23541
  +np.random.seed(s)

   model_opencv8 = mrcal.cameramodel(f"{testdir}/data/cam0.opencv8.cameramodel")
   model_splined = mrcal.cameramodel(f"{testdir}/data/cam0.splined.cameramodel")

Here 23541 is the seed that makes it break. And the time.time() stuff, if
re-enabled, tries out different seeds. The randomness is used in initializing
the solver in mrcal.implied_Rt10__from_unprojections(). The patch here reduces
the size of the numerical fuzz we use when seeding that solver, and I don't see
the test failure anymore

Similarly, the "implied_Rt10 solver moves translation sufficiently. Looking at
difflen at center" test was sometimes failing because too little data was used
to compute the implied transform. I'm using a bit more data now, and the test is
reliably passing. Other seed values that were triggering issues reliably as I
was developing this:

- 23541
- 24560