show_geometry() has smarter logic when given > 2 cameras
Previously I would scale the visualized camera axes with the distance between
the first two cameras we're given. This didn't work well if the first two
cameras were at the same position: the distance, and thus the axes would then
have length 0.

I'm now computing a geometric center of the full set, and basing the axis scale
off that.