Removed range normalization
THIS IS TEMPORARY. Each discrete point once again produces only two
measurements: x,y reprojection error. The third measurement (range
normalization penalty) is gone. I need to do this so that the various
analysis methods (uncertainty) work the same way as with chessboards.
I'm going to re-add this in later, probably in the regularization
terms.