model loading parses yaml/json much more generally
ROS and opencv and kalibr are a mess of inconsistencies. I want a function that
can parse all the various flavors. This doesn't work fully yet: I interpret the
P array in a sensor_msgs/CameraInfo message as some sort of projection matrix,
but it's more like some sort of undistortion array. This needs to ignore it, and
produce identity extrinsics. Sometimes.