Procrustes transform functions detect and report errors
This applies to

  align_procrustes_vectors_R01()
  align_procrustes_points_Rt01()

The input to those functions can be singular (bad data or insufficient data).
Prior to this patch this wasn't detected, and some transform was still being
returned. We now detect this, and report the error as a matrix of 0s. The
results contain a rotation, so an all-0 matrix clearly indicates an error, since
it isn't a valid rotation