Renamed projection uncertainty method 'mean-frames' -> 'mean-pcam'
The math is exactly the same:

  pcam = R_cf pf + t_cf

So if I have many different pf, then mean(R_cf pf + t_cf) = R_cf mean(pf) + t_cf

But formulating it as 'mean-pcam' breaks the requirement to have chessboards at
all. Since the math is the same, the various issues that the old method has are
still there. I should do 'mean-q' or 'best-q' or 'worst-q' instead: these will
be truer representation of what's going on here.

The cross-reprojection method is still preferred, but more implementation, tests
and documentation still need to be written to complete that transition.

This patch renames "mean-frames" to "mean-pcam", and adds some more
documentation to the (renamed) uncertainty-mean-pcam.org file. It does NOT
actually change any code.