Uncertainty guts reimplemented to support moving cameras
This is the first cut. It implements the "mean-pcam" method (the older, existing
method). The implementation isn't completely generic, but generic-enough for
testing. That's ok: I'm doing this for validation only, and this will not be the
final method people will end up using. The goal is to:

- Show that this new implementation produces the same results as the old
  implementation. It should: the logic is the same. If it's different, there's a
  mistake

- Show that stationary-camera and stationary-chessboard uncertainty results are
  identical, if all the data is the same, and we just have a change of reference.
  I don't know for a fact this is the case, but hopefully it is

If both of these can be shown, I can then confidently implement
cross-reprojection using mean-pcam or mean-q or best-q, or something like that