Added mrcal_cameramodel_converter() to libmrcal.so
This is a "converter" function that can be used with "O&" conversions in
PyArg_ParseTupleAndKeywords() calls. Can interpret either path strings or
mrcal.cameramodel objects as mrcal_cameramodel_t structures

This isn't a part of the mrcal Python wrapping, but helps other python
wrapping programs work with mrcal_cameramodel_t. I link this into
libmrcal.so, but libmrcal.so does NOT link with libpython. 99% of the usage
of libmrcal.so will not use this, so it should work without libpython. People
using this function will be doing so as part of
PyArg_ParseTupleAndKeywords(), so they will be linking to libpython anyway.
Thus I weaken all the references to libpython; this is done in the Makefile,
AFTER this is compiled.

This adds a dependency on

- nm           (binutils package)
- llvm-objcopy ("llvm" package)

Probably I just broke cross-building since I'll probably need a special nm for
that