The python-cameramodel-converter is more reliably weakening its symbols
This links this into libmrcal.so, but libmrcal.so does NOT link with libpython.
99% of the usage of libmrcal.so will not use this, so it should work without
libpython. People using this function will be doing so as part of
PyArg_ParseTupleAndKeywords(), so they will be linking to libpython anyway. Thus
I weaken all the references to libpython here inorder to not require linking to
libpython.so for the usages where this is unnecessary

This turns out to be a DEEP rabbithole. With Debian/trixie (released summer
2025) "objcopy --weaken" works to weaken the symbols after compiling. With older
objcopy (not sure which), the objcopy command completes successfully, but it
doesn't actually weaken the symbols. I thus run nm to find the non-weakened
symbols; if any remain, I recompile, explicitly setting __attribute__((weak)) on
these symbols. I can't do this on the first pass because before compiling even
once I don't know which Py symbols I'm going to get (some are generated by the
Python internals).

Finally, the second-pass-recompile-with-attribute-weak doesn't work with
clang: the references produced by inline functions in Python.h that *I* am not
calling are not weakened. I catch this case here, and throw an error. In the
very near future, few people are going to have the too-old binutils, and we'll
be done