python-cameramodel-converter.c doesn't use PyTuple macros
Specifically this patch does PyTuple_GET_ITEM -> PyTuple_GetItem. This is linked
into libmrcal.so WITHOUT creating a strong linkage. So libmrcal.so does NOT have
a NEEDED tag into libpython.so. This is intentional. However this also means
that any ABI details encoded in macros get tied into the code compiled into
libmrcal.so without anything to denote that this is ABI-specific. Here I was
building libmrcal.so with python3.14 and running with python3.13 (or maybe the
other way around...) There was a change regarding the layout of PyTupleObject,
so the PyTuple_GET_ITEM() implementation that was fixed at compile time was now
wrong. Using the function call always does the right thing